✨ Example reveal.js presentation (and npm)
Configures this as an npm package with workspaces. Creates
`presentations` workspace to hold reveal.js markdown-based
presentations. Uses vite to build the presentations into the mdBook
output. Updates GitHub action to orchestrate the build and deploy.

Overall, a hacky-ish approach, but I want to experiment with reveal.js
and host it on my GitHub Pages website.